Task5/llava-v1.5-7b-5GB/requirements.txt
transformers>=4.43.0
accelerate>=0.33.0
bitsandbytes>=0.43.1
sentencepiece
safetensors
huggingface-hub>=0.35.1
pillow>=10.0

